FW: settings: arrayList API rework
There is nothing wrong with the current implementation
it just requires some implicit knowledge about the internals
which should not be used for an API added at such a central
location
To reset the value back to "" you must know that setString
musst be used instead of providing just an empty list which
would be a more "logical" approach
With this the client code would also become easier cause
the check for size()>0 would not be longer required
e.g. in RibbonTargets
I would also suggest to change getArrayList so that it
returns an empty list for the case that the value is ""
instead of an list with size 1 and the single item ""
Allthough this would be a change in the API it would
make the API sync with the putArrayList and would also
be the more "logical" return value for this case
I think I have checked all current uses of getArrayList
and as far as I can see that change would have no impact
on the existing code. But please correct me if I am
wrong with this.

Signed-off-by: igio90 <giovanni.rocca.90@gmail.com>